requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
paho-mqtt==1.4.0
pushbullet.py==0.11.0
python-magic==0.4.15
requests==2.21.0
six==1.12.0
urllib3==1.24.1
websocket-client==0.55.0
